+++ src/solver/heat_equation/homogeneous.cc
@@ -622,7 +622,8 @@ void ConvertMidCylindrical(std::pair<MatrixXd, VectorXd>& mat_b, const Mesh2DReg
     mat_b.first(row, node.u_i_jp) = (1 + r_dash) * mean_xt;
 }
 
-std::pair<MatrixXd, VectorXd> ConvertMesh2dTriangularCartesian(const Mesh2DTriangular& mesh, const VectorXd& results) {
+std::pair<MatrixXd, VectorXd> ConvertMesh2dTriangular(const Mesh2DTriangular& mesh, const VectorXd& results,
+                                                      bool cartesian, bool true_radiation) {
     std::pair<MatrixXd, VectorXd> mat_b;
 
     mat_b.first = MatrixXd::Zero(mesh.nodes_.size(), mesh.nodes_.size());
@@ -632,10 +633,14 @@ std::pair<MatrixXd, VectorXd> ConvertMesh2dTriangularCartesian(const Mesh2DTrian
         const Mesh2DTriangular::Node& node = mesh.nodes_.at(i);
 
         if (node.boundaries.size()) {
-            ConvertBoundariesTriangularMesh(mesh, results, i, mat_b);
-        } else {
+            ConvertBoundariesTriangularMesh(mesh, results, i, mat_b, true_radiation);
+        } else if (cartesian) {
             FEMCentreTriangularMesh(mesh, i, mat_b);
+        } else {
+            FEMCentreCylindricalMesh(mesh, i, mat_b);
         }
+
+        SetVolumetricHeatSourceTriangularMesh(mesh, i, mat_b);
     }
 
     return mat_b;
@@ -841,6 +846,65 @@ void HeatFluxTriangularMesh(const Mesh2DTriangular& mesh, const size_t node_id, 
     }
 }
 
+void RadiationTrueTriangularMesh(const Mesh2DTriangular& mesh, const VectorXd& results, const size_t node_id,
+                                 std::pair<MatrixXd, VectorXd>& mat_b) {
+    const Mesh2DTriangular::Node& node = mesh.nodes_.at(node_id);
+    const Mesh2DTriangular::Boundary& boundary1 = mesh.boundaries_.at(node.boundaries.at(0));
+    const Mesh2DTriangular::Boundary& boundary2 = mesh.boundaries_.at(node.boundaries.at(1));
+    const Vector3d node_pos_left = mesh.GetNodePos(node_id, 0);
+    const Vector3d node_pos_right = mesh.GetNodePos(node_id, node.adjacent_nodes.size() - 1);
+    const double L_left = (node_pos_left - node.position).norm();
+    const double L_right = (node.position - node_pos_right).norm();
+    const double L = L_left + L_right;
+    Vector3d nomal, nomal_left, nomal_right;
+    Matrix3d rot_mat = Matrix3d::Zero();
+    double total_cell_area = 0.0, epsilon;
+
+    rot_mat(0, 1) = 1.0;
+    rot_mat(1, 0) = -1.0;
+    rot_mat(2, 2) = 1.0;
+
+    if (boundary1.type == Mesh2DTriangular::BoundaryType::RADIATION &&
+        boundary2.type == Mesh2DTriangular::BoundaryType::RADIATION) {
+        epsilon = (L_left * boundary1.value + L_right * boundary2.value) / L;
+        nomal_left = (rot_mat * (node_pos_left - node.position)).normalized();
+        nomal_right = (rot_mat * (node.position - node_pos_right)).normalized();
+        nomal = ((L_left * nomal_left + L_right * nomal_right) / L).normalized();
+    } else if (boundary1.type == Mesh2DTriangular::BoundaryType::RADIATION) {
+        nomal = (rot_mat * (node_pos_left - node.position)).normalized();
+        epsilon = boundary1.value;
+    } else {
+        nomal = (rot_mat * (node.position - node_pos_right)).normalized();
+        epsilon = boundary2.value;
+    }
+
+    mat_b.second(node_id) += epsilon * constants::kStefanBoltzmann * std::pow(results(node_id), 4);
+
+    for (size_t c = 0; c < node.adjacent_cells.size(); c++) {
+        const size_t cell_id = node.adjacent_cells.at(c);
+        total_cell_area += mesh.GetCellArea(cell_id);
+    }
+
+    for (size_t c = 0; c < node.adjacent_cells.size(); c++) {
+        const size_t cell_id = node.adjacent_cells.at(c);
+        const Mesh2DTriangular::Cell& cell = mesh.cells_.at(cell_id);
+        const int pos_i = cell.GetNodePos(node_id);
+        const int pos_im = pos_i == 0 ? 2 : pos_i - 1;
+        const int pos_ip = pos_i == 2 ? 0 : pos_i + 1;
+        const size_t node_id_im = cell.nodes.at(pos_im);
+        const size_t node_id_ip = cell.nodes.at(pos_ip);
+        const Mesh2DTriangular::Surface& surface = mesh.surfaces_.at(cell.surface_id);
+        const double frac = mesh.GetCellArea(cell_id) / total_cell_area;
+        const Vector3d coeffs =
+            -frac * surface.thermal_conductivity * CalcNormalDerevativeCoefficients(mesh, nomal, cell_id, node_id);
+
+        mat_b.first(node_id, node_id) += coeffs(0);
+
+        mat_b.first(node_id, node_id_ip) += coeffs(1);
+        mat_b.first(node_id, node_id_im) += coeffs(2);
+    }
+}
+
 void RadiationTriangularMesh(const Mesh2DTriangular& mesh, const VectorXd& results, const size_t node_id,
                              std::pair<MatrixXd, VectorXd>& mat_b) {
     const Mesh2DTriangular::Node& node = mesh.nodes_.at(node_id);
@@ -928,26 +992,6 @@ Vector3d CalcNormalDerevativeCoefficients(const Mesh2DTriangular& mesh, const Ve
     return coeffs;
 }
 
-std::pair<MatrixXd, VectorXd> ConvertMesh2dTriangularCylindrical(const Mesh2DTriangular& mesh,
-                                                                 const VectorXd& results) {
-    std::pair<MatrixXd, VectorXd> mat_b;
-
-    mat_b.first = MatrixXd::Zero(mesh.nodes_.size(), mesh.nodes_.size());
-    mat_b.second = VectorXd::Zero(mesh.nodes_.size());
-
-    for (size_t i = 0; i < mesh.nodes_.size(); i++) {
-        const Mesh2DTriangular::Node& node = mesh.nodes_.at(i);
-
-        if (node.boundaries.size()) {
-            ConvertBoundariesTriangularMesh(mesh, results, i, mat_b);
-        } else {
-            FEMCentreCylindricalMesh(mesh, i, mat_b);
-        }
-    }
-
-    return mat_b;
-}
-
 void FEMCentreCylindricalMesh(const Mesh2DTriangular& mesh, const size_t node_id,
                               std::pair<MatrixXd, VectorXd>& mat_b) {
     const Mesh2DTriangular::Node& node = mesh.nodes_.at(node_id);
@@ -998,7 +1042,7 @@ void FEMCentreCylindricalMesh(const Mesh2DTriangular& mesh, const size_t node_id
 }
 
 void ConvertBoundariesTriangularMesh(const Mesh2DTriangular& mesh, const VectorXd& results, const size_t node_id,
-                                     std::pair<MatrixXd, VectorXd>& mat_b) {
+                                     std::pair<MatrixXd, VectorXd>& mat_b, const bool true_radiation) {
     const Mesh2DTriangular::Node& node = mesh.nodes_.at(node_id);
     const Mesh2DTriangular::Boundary& boundary1 = mesh.boundaries_.at(node.boundaries.at(0));
     const Mesh2DTriangular::Boundary& boundary2 = mesh.boundaries_.at(node.boundaries.at(1));
@@ -1013,6 +1057,9 @@ void ConvertBoundariesTriangularMesh(const Mesh2DTriangular& mesh, const VectorX
     } else if (boundary2.type == Mesh2DTriangular::BoundaryType::DIRICHLET) {
         mat_b.second(node_id) = boundary2.value;
         mat_b.first(node_id, node_id) = 1.0;
+    } else if (true_radiation && (boundary1.type == Mesh2DTriangular::BoundaryType::RADIATION ||
+                                  boundary2.type == Mesh2DTriangular::BoundaryType::RADIATION)) {
+        RadiationTrueTriangularMesh(mesh, results, node_id, mat_b);
     } else if (boundary1.type == Mesh2DTriangular::BoundaryType::RADIATION ||
                boundary2.type == Mesh2DTriangular::BoundaryType::RADIATION) {
         RadiationTriangularMesh(mesh, results, node_id, mat_b);
@@ -1025,6 +1072,22 @@ void ConvertBoundariesTriangularMesh(const Mesh2DTriangular& mesh, const VectorX
     } else {
         throw IncompleteCodeError("undefined boundary condition for triangular mesh");
     }
+}
+
+void SetVolumetricHeatSourceTriangularMesh(const Mesh2DTriangular& mesh, const size_t node_id,
+                                           std::pair<MatrixXd, VectorXd>& mat_b) {
+    const Mesh2DTriangular::Node& node = mesh.nodes_.at(node_id);
+    double volumetric_heat_source = 0.0, total_cell_area = 0.0;
+
+    for (size_t c = 0; c < node.adjacent_cells.size(); c++) {
+        const size_t cell_id = node.adjacent_cells.at(c);
+        const Mesh2DTriangular::Surface surface(mesh.surfaces_.at(mesh.cells_.at(cell_id).surface_id));
+
+        volumetric_heat_source += mesh.GetCellArea(cell_id) * surface.volumetric_heat_source;
+        total_cell_area += mesh.GetCellArea(cell_id);
+    }
+
+    mat_b.second(node_id) += volumetric_heat_source / total_cell_area;
 }
 
 }  // namespace heat_equation_homogeneous