+++ CHANGELOG.md
@@ -6,12 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Algorithms  to solve the 2d homogeneous heat equation on a regular mesh in cartesian coordinates with a volumetric heat source
-- Algorithms  to solve the 2d homogeneous heat equation on a regular mesh in cylinder coordinates with a volumetric heat source
-- Relative error calculation for black body radiation boundary condition
 - Full heat equation on a 2d regular mesh in cartesian coordinates
 - Full heat equation on a 2d regular mesh in cylinder coordinates
 - Writer for hdf5 files
+
+## 2.1.0 - 2024-03-30
+### Added
+- Algorithms to solve the 2d homogeneous heat equation on a triangular mesh in cartesian coordinates with a volumetric heat source
+- Algorithms to solve the 2d homogeneous heat equation on a triangular mesh in cylinder coordinates with a volumetric heat source
+- Relative error calculation for black body radiation boundary condition
 
 ## 2.0.1 - 2024-03-30
 ### Added